[improvement][broker] If there is a deadlock in the service, the probe should return a failure because the service may be unavailable. Add unit testing code.
